pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge"
version = "0.1.0"
description = "Forge — Autonomous CLI Coding Assistant"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "ollama>=0.4.0",
    "mcp>=1.0.0",
    "chromadb>=0.5.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "python-dotenv>=1.0.0",
    "pdfplumber>=0.9.0",
    "trafilatura>=1.6.0",
    "requests>=2.31.0",
]

[project.scripts]
forge = "src.codepilot.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]